fix: test error & now the writing the robust testcases part is left
