+++ v1/Cargo.toml
@@ -10,18 +10,27 @@ crate-type = ["cdylib"]
 
 [dependencies]
 num-rational = "0.4.2"
+
+# Pinocchio crates (always track main)
 pinocchio = { git = "https://github.com/anza-xyz/pinocchio", branch = "main" }
 pinocchio-associated-token-account = { git = "https://github.com/anza-xyz/pinocchio", branch = "main" }
 pinocchio-log = { git = "https://github.com/anza-xyz/pinocchio", branch = "main" }
 pinocchio-pubkey = { git = "https://github.com/anza-xyz/pinocchio", branch = "main" }
 pinocchio-system = { git = "https://github.com/anza-xyz/pinocchio", branch = "main" }
 pinocchio-token = { git = "https://github.com/anza-xyz/pinocchio", branch = "main" }
+
 shank = "0.4.2"
 thiserror = "2.0.12"
 
+# Solana / Agave
+agave-feature-set = "2.3"
+solana-pubkey = "2.4.0"         # bump to resolve conflicts
+solana-program = "2.0.12"
+solana-sdk = "2.0.12"
+
 [dev-dependencies]
-solana-sdk = "1.18"
-solana-program-test = "1.18"
+mollusk-svm = "0.4.2"           # latest mollusk
+solana-program-test = "2.0.12"
 
 [features]
 no-entrypoint = []
